Moved the Playwright media literal slot tests out to their own file.
